+++ src/main/java/com/rental/api/response/GenericResponse.java
